# src/lib/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OrbitABM API",
    "description": "RESTful API for OrbitABM - Account-Based Marketing Campaign Intelligence Platform",
    "version": "1.0.0",
    "contact": {
      "name": "OrbitABM Support",
      "email": "support@orbitabm.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://orbitabm.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/digital-snapshots": {
      "get": {
        "tags": ["Digital Snapshots"],
        "summary": "List digital snapshots",
        "description": "Retrieve a list of digital presence snapshots with optional filtering",
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": true,
            "description": "Organization UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "company_id",
            "in": "query",
            "description": "Filter by company UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "snapshot_date",
            "in": "query",
            "description": "Filter by specific date (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "from_date",
            "in": "query",
            "description": "Filter from date (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "description": "Filter to date (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results per page",
            "schema": {
              "type": "integer",
              "default": 100,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DigitalSnapshotWithCompany"
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Digital Snapshots"],
        "summary": "Create digital snapshot",
        "description": "Create a new digital presence snapshot",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DigitalSnapshotCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Digital snapshot created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DigitalSnapshotWithCompany"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/document-templates/{id}": {
      "get": {
        "tags": ["Document Templates"],
        "summary": "Get document template by ID",
        "description": "Retrieve a single document template by its UUID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Document template UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DocumentTemplate"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "tags": ["Document Templates"],
        "summary": "Update document template",
        "description": "Update an existing document template (partial update)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Document template UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentTemplateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document template updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DocumentTemplate"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Document Templates"],
        "summary": "Delete document template",
        "description": "Soft delete a document template (sets deleted_at timestamp)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Document template UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document template deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "deleted": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DigitalSnapshot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Digital snapshot UUID"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "description": "Organization UUID"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "description": "Company UUID"
          },
          "snapshot_date": {
            "type": "string",
            "format": "date",
            "description": "Snapshot date",
            "example": "2024-02-07"
          },
          "google_rating": {
            "type": "number",
            "minimum": 0,
            "maximum": 5,
            "nullable": true,
            "description": "Google rating (0-5)",
            "example": 4.5
          },
          "google_review_count": {
            "type": "integer",
            "nullable": true,
            "description": "Number of Google reviews",
            "example": 127
          },
          "yelp_rating": {
            "type": "number",
            "minimum": 0,
            "maximum": 5,
            "nullable": true,
            "description": "Yelp rating (0-5)",
            "example": 4.2
          },
          "yelp_review_count": {
            "type": "integer",
            "nullable": true,
            "description": "Number of Yelp reviews",
            "example": 45
          },
          "bbb_rating": {
            "type": "string",
            "nullable": true,
            "description": "Better Business Bureau rating",
            "example": "A+"
          },
          "facebook_followers": {
            "type": "integer",
            "nullable": true,
            "description": "Facebook followers count",
            "example": 1250
          },
          "instagram_followers": {
            "type": "integer",
            "nullable": true,
            "description": "Instagram followers count",
            "example": 890
          },
          "linkedin_followers": {
            "type": "integer",
            "nullable": true,
            "description": "LinkedIn followers count",
            "example": 340
          },
          "domain_authority": {
            "type": "integer",
            "nullable": true,
            "description": "Domain authority score",
            "example": 35
          },
          "page_speed_mobile": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "nullable": true,
            "description": "Mobile page speed score (0-100)",
            "example": 78
          },
          "page_speed_desktop": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "nullable": true,
            "description": "Desktop page speed score (0-100)",
            "example": 85
          },
          "organic_keywords": {
            "type": "integer",
            "nullable": true,
            "description": "Number of organic keywords",
            "example": 245
          },
          "monthly_organic_traffic_est": {
            "type": "integer",
            "nullable": true,
            "description": "Estimated monthly organic traffic",
            "example": 1200
          },
          "website_has_ssl": {
            "type": "boolean",
            "nullable": true,
            "description": "Website has SSL certificate"
          },
          "website_is_mobile_responsive": {
            "type": "boolean",
            "nullable": true,
            "description": "Website is mobile responsive"
          },
          "has_online_booking": {
            "type": "boolean",
            "nullable": true,
            "description": "Has online booking system"
          },
          "has_live_chat": {
            "type": "boolean",
            "nullable": true,
            "description": "Has live chat feature"
          },
          "has_blog": {
            "type": "boolean",
            "nullable": true,
            "description": "Has blog/content section"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Additional notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          }
        },
        "required": ["id", "organization_id", "company_id", "snapshot_date"]
      },
      "DigitalSnapshotWithCompany": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DigitalSnapshot"
          },
          {
            "type": "object",
            "properties": {
              "companies": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "example": "Acme HVAC"
                  },
                  "website": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://acmehvac.com"
                  },
                  "status": {
                    "type": "string",
                    "example": "prospect"
                  },
                  "estimated_revenue": {
                    "type": "number",
                    "nullable": true
                  },
                  "employee_count": {
                    "type": "integer",
                    "nullable": true
                  }
                }
              }
            }
          }
        ]
      },
      "DigitalSnapshotCreate": {
        "type": "object",
        "properties": {
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "description": "Organization UUID"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "description": "Company UUID"
          },
          "snapshot_date": {
            "type": "string",
            "format": "date",
            "description": "Snapshot date (defaults to today)",
            "example": "2024-02-07"
          },
          "google_rating": {
            "type": "number",
            "minimum": 0,
            "maximum": 5,
            "description": "Google rating (0-5)"
          },
          "google_review_count": {
            "type": "integer",
            "description": "Number of Google reviews"
          },
          "yelp_rating": {
            "type": "number",
            "minimum": 0,
            "maximum": 5,
            "description": "Yelp rating (0-5)"
          },
          "yelp_review_count": {
            "type": "integer",
            "description": "Number of Yelp reviews"
          },
          "bbb_rating": {
            "type": "string",
            "description": "Better Business Bureau rating"
          },
          "facebook_followers": {
            "type": "integer",
            "description": "Facebook followers count"
          },
          "instagram_followers": {
            "type": "integer",
            "description": "Instagram followers count"
          },
          "linkedin_followers": {
            "type": "integer",
            "description": "LinkedIn followers count"
          },
          "domain_authority": {
            "type": "integer",
            "description": "Domain authority score"
          },
          "page_speed_mobile": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "description": "Mobile page speed score (0-100)"
          },
          "page_speed_desktop": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "description": "Desktop page speed score (0-100)"
          },
          "organic_keywords": {
            "type": "integer",
            "description": "Number of organic keywords"
          },
          "monthly_organic_traffic_est": {
            "type": "integer",
            "description": "Estimated monthly organic traffic"
          },
          "website_has_ssl": {
            "type": "boolean",
            "description": "Website has SSL certificate"
          },
          "website_is_mobile_responsive": {
            "type": "boolean",
            "description": "Website is mobile responsive"
          },
          "has_online_booking": {
            "type": "boolean",
            "description": "Has online booking system"
          },
          "has_live_chat": {
            "type": "boolean",
            "description": "Has live chat feature"
          },
          "has_blog": {
            "type": "boolean",
            "description": "Has blog/content section"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          }
        },
        "required": ["organization_id", "company_id"]
      },
...
      "DocumentTemplate": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Document template UUID" },
          "organization_id": { "type": "string", "format": "uuid", "description": "Organization UUID" },
          "name": { "type": "string", "description": "Template name", "example": "HVAC Prospect Research" },
          "document_type": { "type": "string", "enum": ["prospect_research", "campaign_sequence", "competitive_analysis", "audit_report", "proposal"], "description": "Document type" },
          "vertical_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Vertical UUID" },
          "template_structure": { "type": "object", "description": "Template structure definition (JSON)" },
          "version": { "type": "integer", "description": "Template version", "example": 1 },
          "is_active": { "type": "boolean", "description": "Whether template is active", "example": true },
          "created_at": { "type": "string", "format": "date-time", "description": "Creation timestamp" },
          "updated_at": { "type": "string", "format": "date-time", "description": "Last update timestamp" },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true, "description": "Deletion timestamp" }
        },
        "required": ["id", "organization_id", "name", "document_type", "template_structure"]
      },
...
      "DocumentTemplateUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Template name" },
          "document_type": { "type": "string", "enum": ["prospect_research", "campaign_sequence", "competitive_analysis", "audit_report", "proposal"], "description": "Document type" },
          "vertical_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Vertical UUID" },
          "template_structure": { "type": "object", "description": "Template structure definition (JSON)" },
          "version": { "type": "integer", "description": "Template version" },
          "is_active": { "type": "boolean", "description": "Whether template is active" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Company with this name already exists"
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "enum": ["VALIDATION_ERROR", "NOT_FOUND", "DUPLICATE_ENTRY", "FORBIDDEN", "DATABASE_ERROR", "SERVER_ERROR", "NETWORK_ERROR"],
            "example": "DUPLICATE_ENTRY"
          },
          "success": {
            "type": "boolean",
            "example": false
          },
          "details": {
            "type": "object",
            "nullable": true,
            "description": "Additional error details",
            "example": {
              "field": "name",
              "value": "Acme HVAC"
            }
          }
        },
        "required": ["error", "code", "success"]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Invalid input data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Organization ID is required",
              "code": "VALIDATION_ERROR",
              "success": false
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found - Resource does not exist",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Company not found",
              "code": "NOT_FOUND",
              "success": false
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict - Resource already exists or constraint violation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Company with this name already exists",
              "code": "DUPLICATE_ENTRY",
              "success": false,
              "details": {
                "field": "name",
                "value": "Acme HVAC"
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal Server Error - Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Internal server error",
              "code": "SERVER_ERROR",
              "success": false
            }
          }
        }
      }
    },
...
  }
}